add policy to company and suppliers / adjustings outhers policies

+++ app/controllers/api/v2/admin/companies_controller.rb
@@ -1,22 +1,24 @@
 class Api::V2::Admin::CompaniesController < Api::V2::Admin::ContactsController
 
   def index
+    companies = policy_scope [:admin, Company] 
     companies = if params['name']
-      Company.where("LOWER(name) LIKE LOWER(?)", "%#{params['name']}%")
+      companies.where("LOWER(name) LIKE LOWER(?)", "%#{params['name']}%")
     else
-      Company.all
+      companies.all
     end
     paginate json: companies.order(:id).as_json(only: [:id, :name, :fantasy_name, :federal_registration]), status: 200
   end
 
   def show
     company = Company.find(params[:id])
+    authorize [:admin, company]
     render json: company, status: 200
   end
 
   def create
     company = Company.new(company_params)
-
+    authorize [:admin, company]
     if company.save
       update_contact(company)
       render json: company, status: 201
@@ -27,6 +29,7 @@ class Api::V2::Admin::CompaniesController < Api::V2::Admin::ContactsController
 
   def update
     company = Company.find(params[:id])
+    authorize [:admin, company]
     if company.update(company_params)
       update_contact(company)
       render json: company, status: 200
@@ -37,6 +40,7 @@ class Api::V2::Admin::CompaniesController < Api::V2::Admin::ContactsController
 
   def destroy
     company = Company.find(params[:id])
+    authorize [:admin, company]
     company.destroy
     head 204
   end
@@ -44,9 +48,6 @@ class Api::V2::Admin::CompaniesController < Api::V2::Admin::ContactsController
   private
 
   def company_params
-    params.permit(:name, :fantasy_name, :federal_registration,
-      :state_registration, :birth_date, :tax_regime, :description,
-      :invoice_sale, :invoice_return, :pis_percent, :confins_percent,
-      :icmsn_percent)
+    params.permit(policy([:admin, Company]).permitted_attributes)
   end
 end
